Make changes to run MNIST (#166)
Steps 2, 5 (only for flattenlayer), 6 completed for issue #165

Big changes for ConvLayer, because it didn't work correctly

---------

+++ include/layers/InputLayer.hpp
@@ -43,7 +43,7 @@ class InputLayer : public Layer {
               "input layer");
         }
         for (int& re : in) {
-          re = static_cast<int>((re - mean_) / std_);
+          re = (re - mean_) / std_;
         }
         Shape sh(input.get_shape());
         if (layin_ == kNchw && layout_ == kNhwc) {
